youtube-ai-platform/python/requirements.txt
mcp==1.10.1
openai-agents==0.1.0
streamlit==1.46.1
youtube-transcript-api==1.1.0
python-dotenv==1.0.0
requests==2.32.4
ipykernel==6.29.5
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76

# LangChain
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
faiss-cpu==1.7.4 